Moved the `string` to `int?` conversion logic into `SettingsNumberBox`
