+++ osu.Game.Tournament/Screens/Editors/RoundEditorScreen.cs
@@ -147,7 +147,7 @@ namespace osu.Game.Tournament.Screens.Editors
                     [Resolved]
                     protected IAPIProvider API { get; private set; }
 
-                    private readonly Bindable<string> beatmapId = new Bindable<string>();
+                    private readonly Bindable<int?> beatmapId = new Bindable<int?>();
 
                     private readonly Bindable<string> mods = new Bindable<string>();
 
@@ -220,14 +220,12 @@ namespace osu.Game.Tournament.Screens.Editors
                     [BackgroundDependencyLoader]
                     private void load(RulesetStore rulesets)
                     {
-                        beatmapId.Value = Model.ID.ToString();
-                        beatmapId.BindValueChanged(idString =>
+                        beatmapId.Value = Model.ID;
+                        beatmapId.BindValueChanged(idInt =>
                         {
-                            int.TryParse(idString.NewValue, out var parsed);
+                            Model.ID = idInt.NewValue ?? 0;
 
-                            Model.ID = parsed;
-
-                            if (idString.NewValue != idString.OldValue)
+                            if (idInt.NewValue != idInt.OldValue)
                                 Model.BeatmapInfo = null;
 
                             if (Model.BeatmapInfo != null)
